define a default scenario first and then add the era-based modifiers on top (address comment slava77)
