+++ Configuration/DataProcessing/python/Impl/ppEra_Run2_2017_trackingOnly.py
@@ -9,13 +9,13 @@ Scenario supporting proton collisions and tracking only reconstruction for HP be
 import os
 import sys
 
-from Configuration.DataProcessing.Reco import Reco
+from   Configuration.DataProcessing.Impl.trackingOnly import trackingOnly
 import FWCore.ParameterSet.Config as cms
-from Configuration.Eras.Era_Run2_2017_cff import Run2_2017
+from   Configuration.Eras.Era_Run2_2017_cff import Run2_2017
 
-from Configuration.DataProcessing.Impl.pp import pp
+from   Configuration.DataProcessing.Impl.pp import pp
 
-class ppEra_Run2_2017_trackingOnly(pp):
+class ppEra_Run2_2017_trackingOnly(trackingOnly):
     def __init__(self):
         pp.__init__(self)
         # tracking only RECO is sufficient, to run high performance BS at PCL;
@@ -25,36 +25,8 @@ class ppEra_Run2_2017_trackingOnly(pp):
         self.eras=Run2_2017
         self.promptCustoms += [ 'Configuration/DataProcessing/RecoTLR.customisePostEra_Run2_2017' ]
         self.expressCustoms += [ 'Configuration/DataProcessing/RecoTLR.customisePostEra_Run2_2017_express_trackingOnly' ]
+        self.alcaHarvCustoms += [ 'Configuration/DataProcessing/RecoTLR.customisePostEra_Run2_2017_harvesting_trackingOnly' ]
         self.visCustoms += [ 'Configuration/DataProcessing/RecoTLR.customisePostEra_Run2_2017' ]
-
-    def expressProcessing(self, globalTag, **args):
-
-        # TkAlMinBias run but hidden to Tier0, in order not to persist it
-        if not args.has_key('skims') :
-            args['skims']=['TkAlMinBias']
-        else :
-            if not 'TkAlMinBias' in args['skims'] :
-                args['skims'].append('TkAlMinBias')
-
-        # reco sequence is limited to tracking => DQM accordingly
-        if not args.has_key('dqmSeq') :
-            args['dqmSeq'] = ['DQMOfflineTracking']
-
-
-        process = pp.expressProcessing(self, globalTag, **args)
-
-        return process
-
-    def alcaHarvesting(self, globalTag, datasetName, **args):
-
-        theCustom = 'Configuration/DataProcessing/RecoTLR.customisePostEra_Run2_2017_harvesting_trackingOnly'
-        if not args.has_key('customs') :
-            args['customs']=[theCustom]
-        else :
-            if not theCustom in args['customs'] :
-                args['customs'].append('TkAlMinBias')
-
-        return pp.alcaHarvesting(self, globalTag, datasetName, **args)
 
     """
     _ppEra_Run2_2017_trackingOnly
